Updated tests, made test helpers immune to coverage reporting
